feat(new): creates new deck with Rofellos as commander
This deck is part of an informal challenge between a friend
and myself - build a deck using a banned commander and with
no banned list.  This deck gets large amounts of value from
Landfall triggers and Rofellos's overpowering mana generation.
Its goal is to use this mana to sink into powerful creatures
and / or value plays.

banned:  only play during this challenge