feat(deck): create casual and normal Jasmine decks
The casual Jasmine Boreal deck is focused on the card
Muraganda Petroglyphs.  All creatures are vanilla and
rely on anthem enchantments to buff them up.  It's
unlikely that this deck can do much, but it seems
fun.

The regular Jasmine Boreal also uses anthem enchant-
ments.  However, it is focused on enchantress effects
for card draw and buffs.