perf(control):  add Daze playset with two Foils
I changed the mana sources to better reflect the
pips involved in most spells.  The largest change
was taking out Arcane Denial in favor of Daze.  I
also threw in a couple copies of Foil which is
fits in with both the budget and enables better
control.  I also added Weather the Storm because
it is a solid addition to any Temur spellslinger.