feat(deck):  create Bant fliers deck
This deck is centered around having fliers
as evasive creatures with some enchantress
effects for value.  The card I am most
looking forward to using is Collective
Voyage.  I think it speeds up the game and
is great for bypassing the required ramping.
My hope is that it'll enable faster games and
more enjoyment so that people can play their
spells.