feat(deck):  create pauper Gruul deck
I wanted to experiment with making a pauper
EDH deck due to the rising cost of MTG.  Even
if this deck does not perform well, the building
process has helped me discover more common cards
that are not often used.  If possible, I want to
start using more commons and uncommons in place
of rares and mythics to help keep costs low.