perf(curve):  remove higher CMCs; replace with lower CMCs
It was sad but I had to remove Ulamog; 10 mana for an
amazing Eldrazi is a splashy play.  However, it does
not speed up my deck or help get me what I need. I also
removed Sensei's Divining Top after testing - as a
friend observed, it forces everyone to spend more time
as I dig for answers.  I added some scry effects and a
copy of Ponder to help improve the mana curve and speed
of the deck.