feat(deck):  create Brion Stoutarm infect deck
This is a casual infect deck that tries to make
Brion have infect so that he can fling creatures
and give players poison counters.  It is fun but
not that viable.  People may feel more pressured
to eliminate the infect player first as poison
creates a timer (10 poison compared to 40 life).