feat(deck):  create Karametra deck
This is a casual Selesnya deck that embodies
the best that Green and White have to offer:
lifegain, counters, going wide, and board
wipes.  I think it is a good choice if we
are sitting down to play with either newer
players or within a more relaxed setting.
There are some amazing drops like The Great
Henge, Savannah, and Nissa, but I think it is
balanced by its small creatures and high CMC
sorceries.