feat(deck):  create throwaway deck with new Omnath
I made this deck based on a thought - "How can I
maximize landfall and elemental tribal?"  I don't
expect to maintain interest in this deck, but it's
fun to play around with.  If I wanted to edit it,
I would remove a few elementals and replace them
with more land focused cards like Oracle of Mul
Daya and Azusa, Lost but Seeking.  The new Thassa
from TBD is also worth considering for her
powerful blink effects.