refactor(decks):  move Niv-Mizzet and Yurlok to Testing
I created a new Testing directory for decks that I have
roughly between a 50-75% interest in physically buying.
It differs from the regular directory in that this is
where I figure out if I enjoy playing deck, am willing
to make modifications to it, and if I think it will be
enjoyable for my playgroup(s).