feat(deck):  create Surrak Dragonclaw deck
This is a commander that I have always loved.
I used to have a deck of him awhile ago, but thought
that he was too slow.  I wanted to give him another
try due to all of the amazing Green and Red cards
that have come out recently.  If I wanted to refine
him, I'd start by looking at the curve and work
backwards.  I want to try and make this version work
so that I have another go-to deck beyond my Rakdos
and Mardu ones.